requirements.txt
streamlit>=1.32.0
groq>=0.9.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
plotly>=5.19.0
requests>=2.31.0